% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindStartStops.R
\name{find_start_and_stop_codons}
\alias{find_start_and_stop_codons}
\title{Find start and stop codons}
\usage{
find_start_and_stop_codons(
  fasta_path,
  start = c("TTG", "CTG", "ATG"),
  stop = c("TAA", "TAG", "TGA"),
  write_bedfiles = TRUE
)
}
\arguments{
\item{fasta_path}{path to a fasta file (DNA) (string)}

\item{start}{a vector of 'start' codons to search for (character)}

\item{stop}{a vector of 'stop' codons to search for (character)}

\item{write_bedfiles}{write bed files describing positions of start/stop codons in each frame (boolean)}
}
\value{
a dataframe describing the positions of start/stop-codons found. Will also write bed files describing start/stop positions if \code{write_bedfiles == true}.
}
\description{
Searches a fasta sequence for all start / stop codons
}
